Rewrote AdminBooksView subpage and its components
