+++ frontend/src/views/AdminTitlesView.vue
@@ -14,6 +14,7 @@
 
 <script lang="ts">
 import { Vue, Component, Prop } from 'vue-property-decorator';
+import TitlesModule from "../store/modules/TitlesModule";
 import TitleFormComponent from '@/components/TitleFormComponent.vue';
 import TitleListComponent from '@/components/TitleListComponent.vue';
 
@@ -23,7 +24,11 @@ import TitleListComponent from '@/components/TitleListComponent.vue';
     TitleListComponent
   }
 })
-export default class AdminTitlesView extends Vue {}
+export default class AdminTitlesView extends Vue {
+  private created() {
+    TitlesModule.fetchAll();
+  }
+}
 </script>
 
 <style scoped>
